CheatEngine: Add a Chromium B.S.U. 0.9.13.2 trainer
I've played around with Cheat Engine 6.4 on Windows to create the
same trainer config which I use for ugtrain and the game Chromium
B.S.U.. On Windows, version 0.9.13.2 is used. I made it to create
the auto-generated game trainer but usability of Cheat Engine is
quite bad for that.