statmemlib: Add a third lib loaded later on
Add a third library which provides a third magic value. Instead of
linking to it, load it with dlopen() after 5 seconds and get the
third magic value by getting its getter function with dlsym() first.
Print the value as soon as the library and the function are
available.
This adds a test for late PIC handling of static memory with ugtrain.